refactor: move traverse flat to abstract array
