--- docarray/array/mixins/__init__.py
@@ -1,7 +1,0 @@
-from docarray.array.mixins.traverse import TraverseMixin
-
-
-class AllMixins(TraverseMixin):
-    """All plugins that can be used in :class:`DocumentArray`."""
-
-    ...
